Combine unlock tasks into one
