+++ game/src/main/kotlin/world/gregs/voidps/world/activity/achievement/TaskSystem.kts
@@ -2,7 +2,6 @@ package world.gregs.voidps.world.activity.achievement
 
 import world.gregs.voidps.engine.client.message
 import world.gregs.voidps.engine.client.ui.chat.plural
-import world.gregs.voidps.engine.client.ui.event.adminCommand
 import world.gregs.voidps.engine.client.ui.event.interfaceOpen
 import world.gregs.voidps.engine.client.ui.interfaceOption
 import world.gregs.voidps.engine.client.ui.open
@@ -218,12 +217,4 @@ interfaceOption("Hint", "hint_*", "task_system") {
     player["world_map_marker_1"] = tile
     player["world_map_marker_text_1"] = ""
     player.open("world_map")
-}
-
-adminCommand("tasks", "auto complete all available tasks", listOf("achievements")) {
-    for (struct in structDefinitions.definitions) {
-        if (struct.stringId.endsWith("_task")) {
-            player[struct.stringId] = true
-        }
-    }
 }
